Conflicts:
	Questor.Modules/Actions/Arm.cs
	Questor.Modules/Actions/BookmarkDestination.cs
	Questor.Modules/Actions/TravelerDestination.cs
	Questor.Modules/Activities/Traveler.cs
	Questor.Modules/BackgroundTasks/Cleanup.cs
	Questor.Modules/BackgroundTasks/NavigateOnGrid.cs
	Questor.Modules/BackgroundTasks/Salvage.cs
	Questor.Modules/Caching/Cache.cs
	Questor.Modules/Combat/Combat.cs
	Questor.Modules/Lookup/Time.cs
	Questor/Behaviors/CombatMissionsBehavior.cs
	Questor/Behaviors/DedicatedBookmarkSalvagerBehavior.cs
	Questor/Questor.cs
	Questor/QuestorUI.cs
	Questor/Storylines/GenericCourierStoryline.cs
	Questor/Storylines/Storyline.cs
	QuestorManager/Actions/ValueDump.cs
	QuestorManager/QuestorManagerUI.cs
	ValueDump/ValueDumpUI.cs

+++ Questor.Modules/Combat/Combat.cs
@@ -151,7 +151,7 @@ namespace Questor.Modules.Combat
             {
                 if (DateTime.Now.Subtract(Cache.Instance.LastLoggingAction).TotalSeconds > 10)
                 {
-                    Cache.Instance.TimeSpentReloading_seconds = Cache.Instance.TimeSpentReloading_seconds + (int)Time.Instance.ReloadWeaponDelayBeforeUsable_seconds;
+                    Cache.Instance.TimeSpentReloading_seconds = Cache.Instance.TimeSpentReloading_seconds + Time.Instance.ReloadWeaponDelayBeforeUsable_seconds;
                     Cache.Instance.LastLoggingAction = DateTime.Now;
                 }
                 Logging.Log("Combat", "Changing [" + weaponNumber + "] with [" + charge.TypeName + "][" + Math.Round((double)ammo.Range / 1000, 0) + "][TypeID: " + charge.TypeId + "] so we can hit [" + entity.Name + "][" + Math.Round(entity.Distance / 1000, 0) + "k]", Logging.teal);
